+++ auth.go
@@ -1,0 +1,96 @@
+package main
+
+import (
+	"encoding/hex"
+	"fmt"
+	"net/http"
+
+	"github.com/gin-gonic/contrib/sessions"
+	"github.com/gin-gonic/gin"
+	"golang.org/x/crypto/bcrypt"
+)
+
+//パスワード処理
+func toHash(password string) (string, error) {
+	hash, err := bcrypt.GenerateFromPassword([]byte(password), 10)
+	if err != nil {
+		return "", err
+	}
+	return hex.EncodeToString(hash), nil
+}
+
+func isTruePassword(password, passwordHash string) bool {
+	decodedPasswordHash, _ := hex.DecodeString(passwordHash)
+	err := bcrypt.CompareHashAndPassword(decodedPasswordHash, []byte(password))
+	if err == nil {
+		//認証に成功した場合にはerrが帰らないので認証成功としてtrueを返す
+		return true
+	}
+	return false
+}
+
+func loginForm(c *gin.Context) {
+	errorMessage, _ := c.Get("loginError")
+	c.HTML(http.StatusOK, "login.tmpl", gin.H{
+		"ErrorMessage": errorMessage,
+	})
+}
+
+func login(c *gin.Context) {
+	username := c.PostForm("username")
+	password := c.PostForm("password")
+	if username == "" || password == "" {
+		//ユーザ・パスが空ならログインに戻す
+		c.Set("loginError", "ユーザ名またはパスワードが間違っています。")
+		loginForm(c)
+	}
+	//ログインチェック
+	isExits, user := isLoginUserExist(username, password)
+	fmt.Printf("userexists%v username:%s", isExits, user.Name)
+	if !isExits {
+		c.Set("loginError", "ユーザ名またはパスワードが間違っています。")
+		loginForm(c)
+	}
+	//セッション作成
+	session := sessions.Default(c)
+
+	session.Set("name", user.Name)
+	session.Set("email", user.Email)
+	session.Set("userId", user.Id)
+	session.Save()
+
+	getTodoList(c)
+
+}
+
+func logout(c *gin.Context) {
+	session := sessions.Default(c)
+	session.Clear()
+	session.Save()
+	loginForm(c)
+}
+
+func isLoginUserExist(username, password string) (bool, User) {
+	var id, name, email, passwordHash string
+	db.QueryRow("SELECT id,name,email,password FROM users WHERE email=$1", username).Scan(&id, &name, &email, &passwordHash)
+	//DBのパスワードと入力されたパスワードをチェック
+	if isTruePassword(password, passwordHash) {
+		//認証成功
+		fmt.Printf("isLoginUserExist認証成功")
+		return true, User{id, name, email}
+	}
+	fmt.Printf("isLoginUserExist認証失敗")
+	return false, User{}
+
+}
+
+//ログインチェック用
+func isLogin(c *gin.Context) {
+	//セッション作成
+	session := sessions.Default(c)
+	userId := session.Get("userId").(string)
+	if userId == "" {
+		//未ログインの場合はログイン画面に飛ばす
+		loginForm(c)
+	}
+}
